refactor(reports): split useReports into focused hooks and colocate data logic in cards
