+++ client/src/components/pages/Store/Reports/hooks/__tests__/useReports.test.js
@@ -2,7 +2,7 @@ import { act, renderHook } from "@testing-library/react";
 
 import { httpClient, parseJsonResponse } from "@/lib/http";
 
-import { defaultFilters, useDateRangeFilters, useReports } from "../useReports";
+import { useReports } from "../useReports";
 
 jest.mock("@/lib/http", () => ({
   httpClient: jest.fn(),
@@ -280,332 +280,5 @@ describe("useReports — fetch", () => {
 
     const url = httpClient.mock.calls[0][0];
     expect(url).not.toContain("userId");
-  });
-});
-
-describe("useReports — filters", () => {
-  beforeEach(() => {
-    jest.clearAllMocks();
-    httpClient.mockResolvedValue({});
-    parseJsonResponse.mockResolvedValue(successReport);
-  });
-
-  it("initial filters match defaultFilters", async () => {
-    const { result } = renderHook(() => useReports());
-    await act(async () => {});
-    expect(result.current.filters).toEqual(defaultFilters);
-  });
-
-  it("auto-fetches with default period on mount", async () => {
-    renderHook(() => useReports());
-    await act(async () => {});
-    expect(httpClient).toHaveBeenCalledWith(
-      expect.stringContaining(`period=${defaultFilters.activePeriod}`),
-    );
-  });
-
-  it("handleFiltersChange with activePeriod fetches immediately", async () => {
-    const { result } = await setupHook();
-
-    await act(async () => {
-      result.current.handleFiltersChange({ activePeriod: "week", startDate: "", endDate: "" });
-    });
-
-    expect(httpClient).toHaveBeenCalledWith(expect.stringContaining("period=week"));
-  });
-
-  it("handleFiltersChange with activePeriod updates filters state", async () => {
-    const { result } = await setupHook();
-
-    await act(async () => {
-      result.current.handleFiltersChange({ activePeriod: "year", startDate: "", endDate: "" });
-    });
-
-    expect(result.current.filters.activePeriod).toBe("year");
-  });
-
-  it("handleFiltersChange with only startDate does not fetch", async () => {
-    const { result } = await setupHook();
-
-    act(() => {
-      result.current.handleFiltersChange({ startDate: "2024-01-01", activePeriod: null });
-    });
-
-    expect(httpClient).not.toHaveBeenCalled();
-  });
-
-  it("handleFiltersChange with only endDate does not fetch", async () => {
-    const { result } = await setupHook();
-
-    act(() => {
-      result.current.handleFiltersChange({ endDate: "2024-12-31", activePeriod: null });
-    });
-
-    expect(httpClient).not.toHaveBeenCalled();
-  });
-
-  it("handleFiltersChange fetches when both startDate and endDate are valid", async () => {
-    const { result } = await setupHook();
-
-    await act(async () => {
-      result.current.handleFiltersChange({ startDate: "2024-01-01", activePeriod: null });
-    });
-    await act(async () => {
-      result.current.handleFiltersChange({ endDate: "2024-01-31", activePeriod: null });
-    });
-
-    const urls = httpClient.mock.calls.map((c) => c[0]);
-    expect(urls.some((u) => u.includes("startDate=2024-01-01"))).toBe(true);
-    expect(urls.some((u) => u.includes("endDate=2024-01-31"))).toBe(true);
-  });
-
-  it("handleFiltersChange with paymentMethod fetches immediately", async () => {
-    const { result } = await setupHook();
-
-    await act(async () => {
-      result.current.handleFiltersChange({ paymentMethod: "Cash" });
-    });
-
-    expect(httpClient).toHaveBeenCalledWith(
-      expect.stringContaining("paymentMethod=Cash"),
-    );
-  });
-
-  it("handleFiltersChange with productName debounces 500ms", async () => {
-    jest.useFakeTimers();
-    const { result } = renderHook(() => useReports());
-    await act(async () => { jest.runAllTimers(); });
-    jest.clearAllMocks();
-
-    act(() => {
-      result.current.handleFiltersChange({ productName: "Widget" });
-    });
-    expect(httpClient).not.toHaveBeenCalled();
-
-    act(() => { jest.advanceTimersByTime(500); });
-    await act(async () => {});
-
-    expect(httpClient).toHaveBeenCalledWith(
-      expect.stringContaining("productName=Widget"),
-    );
-
-    jest.useRealTimers();
-  });
-
-  it("generateReport fetches with current filters", async () => {
-    const { result } = await setupHook();
-
-    await act(async () => {
-      await result.current.generateReport();
-    });
-
-    expect(httpClient).toHaveBeenCalledWith(
-      expect.stringContaining(`period=${defaultFilters.activePeriod}`),
-    );
-  });
-});
-
-describe("useReports — derived values", () => {
-  beforeEach(() => {
-    jest.clearAllMocks();
-    httpClient.mockResolvedValue({});
-  });
-
-  it("totalRevenue is 0 when reportData is null", async () => {
-    const { result } = renderHook(() => useReports());
-    await act(async () => {});
-    expect(result.current.totalRevenue).toBe(0);
-  });
-
-  it("totalItems is 0 when reportData is null", async () => {
-    const { result } = renderHook(() => useReports());
-    await act(async () => {});
-    expect(result.current.totalItems).toBe(0);
-  });
-
-  it("totalRevenue derives from reportData.totalRevenueCents", async () => {
-    parseJsonResponse.mockResolvedValue({ totalRevenueCents: 9900, totalItemsSold: 0, sales: [] });
-    const { result } = await setupHook();
-    parseJsonResponse.mockResolvedValueOnce({ totalRevenueCents: 9900, totalItemsSold: 0, sales: [] });
-
-    await act(async () => {
-      await result.current.fetchReport({ period: "month" });
-    });
-
-    expect(result.current.totalRevenue).toBe(9900);
-  });
-
-  it("totalItems derives from reportData.totalItemsSold", async () => {
-    parseJsonResponse.mockResolvedValue({ totalRevenueCents: 0, totalItemsSold: 42, sales: [] });
-    const { result } = await setupHook();
-    parseJsonResponse.mockResolvedValueOnce({ totalRevenueCents: 0, totalItemsSold: 42, sales: [] });
-
-    await act(async () => {
-      await result.current.fetchReport({ period: "month" });
-    });
-
-    expect(result.current.totalItems).toBe(42);
-  });
-});
-
-describe("useDateRangeFilters", () => {
-  const baseFilters = { activePeriod: "", startDate: "", endDate: "", productName: "", paymentMethod: "" };
-
-  it("activeFilterCount counts non-empty filter values", () => {
-    const { result } = renderHook(() => useDateRangeFilters({ ...baseFilters, activePeriod: "month", productName: "cake" }, jest.fn()),
-    );
-    expect(result.current.activeFilterCount).toBe(2);
-  });
-
-  it("dateRangeValue is null when dates are missing", () => {
-    const { result } = renderHook(() => useDateRangeFilters(baseFilters, jest.fn()));
-    expect(result.current.dateRangeValue).toBeNull();
-  });
-
-  it("dateRangeValue returns CalendarDate objects when both dates are set", () => {
-    const { result } = renderHook(() => useDateRangeFilters({ ...baseFilters, startDate: "2024-01-01", endDate: "2024-01-31" }, jest.fn()),
-    );
-    expect(result.current.dateRangeValue.start.toString()).toBe("2024-01-01");
-    expect(result.current.dateRangeValue.end.toString()).toBe("2024-01-31");
-  });
-
-  it("handlePeriodChange calls onFiltersChange with period and clears dates", () => {
-    const onFiltersChange = jest.fn();
-    const { result } = renderHook(() => useDateRangeFilters(baseFilters, onFiltersChange));
-    result.current.handlePeriodChange("week");
-    expect(onFiltersChange).toHaveBeenCalledWith({ activePeriod: "week", startDate: "", endDate: "" });
-  });
-
-  it("handleDateRangeChange with a range calls onFiltersChange with stringified dates", () => {
-    const onFiltersChange = jest.fn();
-    const { result } = renderHook(() => useDateRangeFilters(baseFilters, onFiltersChange));
-    result.current.handleDateRangeChange({
-      start: { toString: () => "2024-03-01" },
-      end: { toString: () => "2024-03-31" },
-    });
-    expect(onFiltersChange).toHaveBeenCalledWith({ startDate: "2024-03-01", endDate: "2024-03-31", activePeriod: null });
-  });
-
-  it("handleDateRangeChange with null clears dates and activePeriod", () => {
-    const onFiltersChange = jest.fn();
-    const { result } = renderHook(() => useDateRangeFilters(baseFilters, onFiltersChange));
-    result.current.handleDateRangeChange(null);
-    expect(onFiltersChange).toHaveBeenCalledWith({ startDate: "", endDate: "", activePeriod: null });
-  });
-
-  it("handlePaymentMethod with 'all' calls onFiltersChange with empty string", () => {
-    const onFiltersChange = jest.fn();
-    const { result } = renderHook(() => useDateRangeFilters(baseFilters, onFiltersChange));
-    result.current.handlePaymentMethod(new Set(["all"]));
-    expect(onFiltersChange).toHaveBeenCalledWith({ paymentMethod: "" });
-  });
-
-  it("handlePaymentMethod with a specific method passes it through", () => {
-    const onFiltersChange = jest.fn();
-    const { result } = renderHook(() => useDateRangeFilters(baseFilters, onFiltersChange));
-    result.current.handlePaymentMethod(new Set(["BTC"]));
-    expect(onFiltersChange).toHaveBeenCalledWith({ paymentMethod: "BTC" });
-  });
-});
-
-describe("useReports — computed analytics", () => {
-  const salesReport = {
-    totalRevenueCents: 7000,
-    totalItemsSold: 5,
-    sales: [
-      { productName: "A", quantity: 2, priceAtOrder: 1000, userName: "u1", paymentMethod: "Cash", saleDate: "2024-01-02T10:00:00" },
-      { productName: "B", quantity: 1, priceAtOrder: 3000, userName: "u2", paymentMethod: "BTC", saleDate: "2024-01-01T09:00:00" },
-      { productName: "A", quantity: 1, priceAtOrder: 1000, userName: "u1", paymentMethod: "Cash", saleDate: "2024-01-02T14:00:00" },
-    ],
-  };
-
-  beforeEach(() => {
-    jest.clearAllMocks();
-    httpClient.mockResolvedValue({});
-  });
-
-  async function setupWithSales() {
-    const hook = renderHook(() => useReports());
-    await act(async () => {});
-    jest.clearAllMocks();
-    httpClient.mockResolvedValue({});
-    parseJsonResponse.mockResolvedValueOnce(salesReport);
-    await act(async () => { await hook.result.current.fetchReport({ period: "month" }); });
-    return hook;
-  }
-
-  it("sales is [] when reportData is null", () => {
-    parseJsonResponse.mockResolvedValue({ totalRevenueCents: 0, totalItemsSold: 0, sales: [] });
-    const { result } = renderHook(() => useReports());
-    expect(result.current.sales).toEqual([]);
-  });
-
-  it("sales derives from reportData.sales", async () => {
-    const { result } = await setupWithSales();
-    expect(result.current.sales).toEqual(salesReport.sales);
-  });
-
-  it("revenueByDay groups sales by date and sums quantity * priceAtOrder", async () => {
-    const { result } = await setupWithSales();
-    const jan2 = result.current.revenueByDay.find((d) => d.date === "2024-01-02");
-    expect(jan2.revenue).toBe(3000); // 2*1000 + 1*1000
-  });
-
-  it("revenueByDay is sorted ascending by date", async () => {
-    const { result } = await setupWithSales();
-    const dates = result.current.revenueByDay.map((d) => d.date);
-    expect(dates).toEqual([...dates].sort());
-  });
-
-  it("revenueByDay count tracks total quantity per day", async () => {
-    const { result } = await setupWithSales();
-    const jan2 = result.current.revenueByDay.find((d) => d.date === "2024-01-02");
-    expect(jan2.count).toBe(3); // quantity 2 + 1
-  });
-
-  it("topProducts sorts by revenue descending", async () => {
-    const { result } = await setupWithSales();
-    const revenues = result.current.topProducts.map((p) => p.revenue);
-    expect(revenues).toEqual([...revenues].sort((a, b) => b - a));
-  });
-
-  it("topProducts aggregates quantity across multiple sales of the same product", async () => {
-    const { result } = await setupWithSales();
-    const productA = result.current.topProducts.find((p) => p.name === "A");
-    expect(productA.quantity).toBe(3);
-  });
-
-  it("topProducts is capped at 8 entries", async () => {
-    const manySales = Array.from({ length: 10 }, (_, idx) => ({
-      productName: `Product ${idx}`,
-      quantity: 1,
-      priceAtOrder: 1000 - idx * 10,
-      userName: "u",
-      paymentMethod: "Cash",
-      saleDate: "2024-01-01T00:00:00",
-    }));
-    parseJsonResponse.mockResolvedValue({ totalRevenueCents: 0, totalItemsSold: 0, sales: [] });
-    const hook = renderHook(() => useReports());
-    await act(async () => {});
-    jest.clearAllMocks();
-    httpClient.mockResolvedValue({});
-    parseJsonResponse.mockResolvedValueOnce({ totalRevenueCents: 0, totalItemsSold: 0, sales: manySales });
-    await act(async () => { await hook.result.current.fetchReport({ period: "month" }); });
-    expect(hook.result.current.topProducts).toHaveLength(8);
-  });
-
-  it("paymentMethodSplit groups by paymentMethod and sorts by revenue descending", async () => {
-    const { result } = await setupWithSales();
-    const [first] = result.current.paymentMethodSplit;
-    expect(first.method).toBe("Cash"); // Cash: 3000, BTC: 3000 — Cash has higher count but same revenue; order is stable
-    result.current.paymentMethodSplit.forEach((entry, idx, arr) => {
-      if (idx > 0) expect(entry.revenue).toBeLessThanOrEqual(arr[idx - 1].revenue);
-    });
-  });
-
-  it("paymentMethodSplit count tracks number of sale records per method", async () => {
-    const { result } = await setupWithSales();
-    const cash = result.current.paymentMethodSplit.find((e) => e.method === "Cash");
-    expect(cash.count).toBe(2);
   });
 });
