+++ client/src/components/pages/Store/Reports/hooks/useSummaryData.js
@@ -1,0 +1,8 @@
+"use client";
+import { useMemo } from "react";
+
+export function useSummaryData(reportData) {
+  const totalRevenue = useMemo(() => reportData?.totalRevenueCents ?? 0, [reportData]);
+  const totalItems = useMemo(() => reportData?.totalItemsSold ?? 0, [reportData]);
+  return { totalRevenue, totalItems };
+}
